normalize routing
