+++ app/views/orders/new.html.erb
@@ -1,13 +1,11 @@
 <div class="col-sm-12 col-md-12 col-ld-6 col-ld-offset-3">
-  <%= simple_form_for [:account, @order] do |f|%>
+  <%= simple_form_for [:plan, @order] do |f|%>
     <div class="form-group">
       <%= f.input :backer_name, input_html:{class: "form-control"}, placeholder: "请填写您的姓名" %>
       <%= f.input :price, input_html: {class: "form-control", :readonly => true, value: @order.price }, placeholder: "请填写价格"%>
       <%= f.input :quantity, input_html: {class: "form-control", :readonly => true, value: @order.quantity },  placeholder: "请填写数量" %>
-      <%= f.input :plan_id, :as => :hidden, :input_html => { :value => @order.plan_id } %>
       <%= f.submit "Submit", class: "btn btn-primary", data: {disable_with: "Submitting.."} %>
     </div>
-
   <% end %>
       <!-- <div style="margin-top: 50px;">
           <h4>请扫下面的二维码进行付款</h4>
